Xbox Controls
Left Joystick for movement. If you click it down you sprint.
Right Joystick for camera. If you click it down you roll.
Right trigger is a light attack.
Left trigger is a heavy attack.
A to jump.
B to enter the build menu.
In the build menu:
Right trigger to build.
DPad left and right to select buildings.
Left and right bumpers to rotate building before placing.
Start to pause.